Deduplicate `JsFuture` definitions
Turns out it's the exact same for both before and after atomics, so
let's use the same definition!
